19.05.2021

+++ src/main/java/ru/izotov/userphonebooks/services/BookEntryService.java
@@ -11,11 +11,9 @@ import ru.izotov.userphonebooks.models.PhoneBook;
 import ru.izotov.userphonebooks.repositories.PhoneBookRepo;
 import ru.izotov.userphonebooks.repositories.BookEntryRepo;
 import ru.izotov.userphonebooks.repositories.UserRepo;
-import ru.izotov.userphonebooks.services.utils.BookEntryUtils;
 import ru.izotov.userphonebooks.services.utils.UserUtils;
 
 import java.util.*;
-import java.util.stream.Collectors;
 
 @Service
 public class BookEntryService {
@@ -29,47 +27,30 @@ public class BookEntryService {
 
     @Autowired
     private UserUtils userUtils;
-    @Autowired
-    private BookEntryUtils entryUtils;
 
 
     public PhoneBook getAllEntriesForUser(Long user_id) throws Exception {
-        UserEntity userEntity = userRepo
-                .findById(user_id)
-                .orElseThrow(() -> new UserInteractionException(String.format("User with id %d not found", user_id)));
+        UserEntity userEntity = userUtils.findByIdOrThrows(user_id);
+        List<PhoneBookEntity> book = userEntity.getBookEntities();
 
-        List<PhoneBookEntity> entries =  bookRepo.findByOwner_id(user_id);
-
-        if(entries.isEmpty()){
+        if(book.isEmpty()){
             throw new UserInteractionException(String.format("User %s has no entries in the phone book", userEntity.getUserName()));
         }
 
-        return PhoneBook.toModel(entries);
+        return PhoneBook.toModel(book);
     }
 
     public Long delete(Long id, Long entry_id) throws UserInteractionException {
         UserEntity user = userUtils.findByIdOrThrows(id);
 
-        PhoneBookEntity entry = user.getBookEntities().stream()
-                .filter(b -> b.getId().equals(entry_id))
-                .findFirst()
-                .orElseThrow(()->new UserInteractionException(String.format("Entry with id %d not found", entry_id)));
-        bookRepo.findById(entry_id);
+        if(user.getBookEntities().stream().noneMatch(b -> b.getId() == entry_id)){
+            throw new UserInteractionException(String.format("Entry with id %d not found", entry_id));
+        }
         bookRepo.deleteById(entry_id);
         return entry_id;
     }
 
 
-    public PhoneBook findBiId(Long user_id, Long entry_id) throws UserInteractionException {
-        UserEntity user = userUtils.findByIdOrThrows(user_id);
-        List<PhoneBookEntity> book = user.getBookEntities();
-        PhoneBookEntity entry = book.stream()
-                .filter(b->b.getId().equals(entry_id))
-                .findFirst()
-                .orElseThrow(()->new UserInteractionException(String.format("User %s does not have an entry with id %s", user.getUserName(), entry_id)));
-        return PhoneBook.toModel(entry);
-
-    }
 
     public PhoneBook findByPhoneNumber(Long user_id, String phoneNumber) throws UserInteractionException {
         UserEntity user = userUtils.findByIdOrThrows(user_id);
@@ -86,88 +67,82 @@ public class BookEntryService {
     public BookEntry editEntry(BookEntryEntity editEntry, Long entry_id, Long user_id) throws UserInteractionException {
         UserEntity user = userUtils.findByIdOrThrows(user_id);
 
-        PhoneBookEntity bookEntry = user.getBookEntities().stream()
-                .filter(entry -> entry.getId().equals(entry_id))
+        BookEntryEntity entry = user.getBookEntities().stream()
+                .filter(entry1 -> entry1.getId().equals(entry_id))
                 .findFirst()
-                .orElseThrow(()->new UserInteractionException(String.format("No entry with id %s found", entry_id)));
+                .orElseThrow(()->new UserInteractionException(String.format("No entry with id %s found", entry_id)))
+                .getEntry();
 
-        Optional<String> userName = Optional.ofNullable(editEntry.getUserName());
-        if(userName.isPresent() && !userName.get().isEmpty()){
-            bookEntry.getEntry().setUserName(userName.get());
-            UserEntity userPhone = bookEntry.getEntry().getUser();
-            if(userPhone!=null) userPhone.setUserName(userName.get());
-        }
+        Optional.ofNullable(editEntry.getUserName())
+                .ifPresent(name -> {
+            if(!name.isEmpty()
+                    && name.length() < 50){
+                entry.setUser(null);
+                entry.setUserName(name);
+            }
+        });
 
         Optional<String> phoneNumber = Optional.ofNullable(editEntry.getPhoneNumber());
         if(phoneNumber.isPresent() && !(phoneNumber.get().length() > 6)){
-            bookEntry.getEntry().setPhoneNumber(phoneNumber.get());
+            entry.setPhoneNumber(phoneNumber.get());
         }
-        bookRepo.save(bookEntry);
-        return BookEntry.toModel(bookEntry.getEntry());
+
+        Optional.ofNullable(editEntry.getPhoneNumber())
+                .ifPresent(number ->{
+            if(!number.isEmpty()
+                    && number.length() > 6
+                    && number.length() < 30){
+                entry.setPhoneNumber(number);
+            }
+        });
+        entryRepo.save(entry);
+        return BookEntry.toModel(entry);
     }
 
-    //По возможности переделать
-    public PhoneBook createEntry(BookEntryEntity newEntry, Long user_id) throws UserInteractionException {
+    public BookEntry createEntry(BookEntryEntity newEntry, Long user_id) throws UserInteractionException {
         UserEntity owner = userUtils.findByIdOrThrows(user_id);
-        if(!entryUtils.isUserName(newEntry.getUserName()) || !entryUtils.isPhoneNumber(newEntry.getPhoneNumber())){
-            throw new UserInteractionException("The new entry contains an invalid username and / or password");
-        }
-        newEntry = entryUtils.findByUsernameAndPhoneNumberOrThrows(newEntry);
 
+        String userName = newEntry.getUserName();
         String phoneNumber = newEntry.getPhoneNumber();
-        List<PhoneBookEntity> book = owner.getBookEntities();
-        if(book.stream().anyMatch(b -> b.getEntry().getPhoneNumber() == phoneNumber)){
+
+        if(owner.getBookEntities().stream().anyMatch(b -> b.getEntry().getPhoneNumber() == phoneNumber)){
             throw new UserInteractionException("An entry with this number already exists");
         }
-        //numberRepo.save(newEntry);
+
+        if(userName==null || userName.isEmpty()
+                || userName.length() > 50){
+            throw new UserInteractionException("The new entry contains an invalid user name");
+        }
+
+        if(phoneNumber==null || phoneNumber.isEmpty()
+                || phoneNumber.length() < 4 || phoneNumber.length() > 30){
+            throw new UserInteractionException("The new entry contains an invalid phone number");
+        }
+
+        BookEntryEntity dbEntry = entryRepo.findByPhoneNumber(phoneNumber).orElse(newEntry);
+
+        /* Если запись не найдена по номеру телефона -> ищем совпадение по имени
+         * в таблице User и присваеваем его если нашли
+         * * */
+        if(dbEntry == newEntry){
+            UserEntity user;
+            if((user = userRepo.findByUserName(dbEntry.getUserName())) !=null ){
+                dbEntry.setUser(user);
+            }
+            /* Сравниваем имена у двух записей:
+             *       1. Имена совпали -> вернули запись из базы
+             *       2. Имена разные -> бросили исключение
+             * */
+        }else if(!dbEntry.getUserName().equals(newEntry.getUserName())){
+            throw new UserInteractionException("A user with this phone number already exists");
+        }
+
+        //newEntry = entryUtils.findByUsernameAndPhoneNumberOrThrows(newEntry);
+        dbEntry = entryRepo.save(dbEntry);
         PhoneBookEntity newBookEntry = new PhoneBookEntity();
         newBookEntry.setOwner(owner);
-        newBookEntry.setEntry(newEntry);
+        newBookEntry.setEntry(dbEntry);
         bookRepo.save(newBookEntry);
-        return PhoneBook.toModel(newBookEntry);
-
-
-        /*
-        //Проверка имени пользователя на соответствие
-        String username = checkingUsername(newEntry.getUsername());
-
-        //Проверка телефонного номера на соответствие
-        String phonenumber = checkingPhonenumber(newEntry.getPhonenumber());
-
-        Optional<PhoneNumberEntity> entry;
-        if((entry = numberRepo.findByPhonenumber(phonenumber)).isPresent()){
-            if(!entry.get().getUsername().equals(username)){
-                //У двух разный username не может быть одинаковых телефонных номеров
-                throw new FieldAlreadyExistException("A user with this phone number already exists");
-            }
-            newEntry = entry.get();
-        }
-
-        //Проверка владельца телефонной книги на присутствие в базе
-        UserEntity owner = findUser(user_id);
-
-
-        //Ищет пользователя с таким именем и присваивает ему номер в случае его нахождения в базе
-        Optional<UserEntity> user;
-        if(newEntry.getUser() == null && (user= userRepo.findByUsername(username)).isPresent()){
-            newEntry.setUser(user.get());
-        }
-
-        if(bookRepo
-                .findByUser_id(owner.getId())
-                .stream()
-                .noneMatch(rec -> rec.getNumber().getPhonenumber().equals(phonenumber))){
-
-            numberRepo.save(newEntry);
-            //Создается запись и сохраняется в телефонную книгу
-            PhoneBookEntity bookEntity = new PhoneBookEntity();
-            bookEntity.setUser(owner);
-            bookEntity.setNumber(newEntry);
-            bookRepo.save(bookEntity);
-        }else {
-            throw new FieldAlreadyExistException("This entry already exists in the phone book");
-        }
-
-        return PhoneNumber.toModel(newEntry);*/
+        return BookEntry.toModel(dbEntry);
     }
 }
